requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
scikit-learn==1.3.2
tensorflow==2.14.0
paho-mqtt==1.6.1
websockets==12.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
schedule==1.2.0
redis==5.0.1
psutil==5.9.6
seaborn==0.13.0
matplotlib==3.8.2
scipy==1.11.4
joblib==1.3.2
watchdog==3.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
